+++ recipes/default.rb
@@ -1,10 +1,10 @@
 #
-# Cookbook Name:: ad-nativex
+# Cookbook Name:: ad-simplyadrian
 # Recipe:: default 
 #
-# Copyright 2014, NativeX
+# Copyright 2014, simplyadrian
 #
 # All rights reserved - Do Not Redistribute
 #
 
-include_recipe 'ad-nativex::joindomain'
+include_recipe 'ad-simplyadrian::joindomain'