+++ recipes/joindomain.rb
@@ -1,16 +1,16 @@
 #
-# Cookbook Name:: ad-nativex
+# Cookbook Name:: ad-simplyadrian
 # Recipe:: joindomain
 #
-# Copyright 2014, NativeX
+# Copyright 2014, simplyadrian
 #
 # All rights reserved - Do Not Redistribute
 #
 
 # If we're in EC2, then need to dynamically determine the the OU based on region
-include_recipe 'ad-nativex::dynamic_ou'
+include_recipe 'ad-simplyadrian::dynamic_ou'
 creds = Chef::EncryptedDataBagItem.load("credentials", "ad")
-domain = node['ad-nativex']['name']
+domain = node['ad-simplyadrian']['name']
 
 if centos?
 
@@ -20,7 +20,7 @@ if centos?
 
     package 'adcli'
 
-    include_recipe 'ad-nativex::dynamic_dc'
+    include_recipe 'ad-simplyadrian::dynamic_dc'
 
     # Configure Kerberos
     template '/etc/krb5.conf' do
@@ -35,7 +35,7 @@ if centos?
         if domain_info.include? domain
           cmd = "echo -n #{creds['ad_password']} | adcli join --domain=#{domain} "\
             "--login-user=#{creds['ad_username'].split('@')[0]}@#{domain.upcase} "\
-            "--stdin-password --domain-ou=\"#{node['ad-nativex']['oupath']}\" --show-details"
+            "--stdin-password --domain-ou=\"#{node['ad-simplyadrian']['oupath']}\" --show-details"
           join_domain = `#{cmd}`
           Chef::Log.info(join_domain)
         else
@@ -54,7 +54,7 @@ if centos?
     end
 
     # Configure SSSD
-    include_recipe 'ad-nativex::sssd_ldap'
+    include_recipe 'ad-simplyadrian::sssd_ldap'
 
   end
 
@@ -64,12 +64,12 @@ elsif windows?
   include_recipe 'windows::reboot_handler'
   node.default[:windows][:allow_pending_reboots] = false
 
-  # Join teamfreeze.com domain
-  ad_nativex_domain "#{node['ad-nativex']['name']}" do
+  # Join defaultdomain.com domain
+  ad_simplyadrian_domain "#{node['ad-simplyadrian']['name']}" do
     action :join
     domain_pass creds["ad_password"]
     domain_user creds["ad_username"]
-    oupath lazy { "\"#{node['ad-nativex']['oupath']}\"" }
+    oupath lazy { "\"#{node['ad-simplyadrian']['oupath']}\"" }
     notifies :request, 'windows_reboot[60]', :delayed
   end