+++ recipes/promotetodomaincontroller.rb
@@ -1,8 +1,8 @@
 #
-# Cookbook Name:: ad-nativex
+# Cookbook Name:: ad-simplyadrian
 # Recipe:: promotetodomaincontroller
 #
-# Copyright 2014, NativeX
+# Copyright 2014, simplyadrian
 #
 # All rights reserved - Do Not Redistribute
 #
@@ -11,12 +11,12 @@
 
 # Promote to Domain Controller
 creds = Chef::EncryptedDataBagItem.load("credentials", "ad")
-ad_nativex_domaincontroller "#{node['ad-nativex']['name']}" do
+ad_simplyadrian_domaincontroller "#{node['ad-simplyadrian']['name']}" do
   action :create
   type "replica" 
   domain_pass creds["ad_password"]
   domain_user creds["ad_username"]
-  site_name "#{node['ad-nativex']['site_name']}"
-  safe_mode_pass "#{node['ad-nativex']['safe_mode_pass']}"
+  site_name "#{node['ad-simplyadrian']['site_name']}"
+  safe_mode_pass "#{node['ad-simplyadrian']['safe_mode_pass']}"
 end